Update patch set 2
Patch Set 2:

If this is the only change, then the problem is most likely that the module is registered in the system on the server, but never associated with any page and never loaded. So it would not do anything as of yet.

Patch-set: 2
Reviewer: Krinkle <34@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0