Update patch set 9
Patch Set 9:

Ah ok, didn't see that for some reason. It's to join directly then… but I'm not sure about the performance implications (would need to look into that some more).

Also as said before, the script should batch itself, currently you will need to run it in a loop, which is not very nice.

Patch-set: 9