Update patch set 2
Patch Set 2:

(2 comments)

It seems this will collect an aweful lot of information, all in memory. I wonder if there is no way to let the SQL server let the majority of the work done in here, and most importantly let it work in chunks right from the start.

The relevant part is the "HAVING COUNT(*) > 1" condition. This only returns stuff that is guaranteed to have at least one duplicate.

Deleting the actual duplicates should be possible with something like WHERE id <> the one id that was left after the GROUP BY.

Patch-set: 2