Update patch set 2
Patch Set 2:

(4 comments)

Why is this merged? I'm sorry that my comments are not understood. I will try again. To summarize again: This script is broken and can not work for multiple reasons. In a best case scenario it will just do nothing. What's most probably is that is will delete some duplicates but not others. In a worst case scenario it will delete wrong stuff.

Patch-set: 2