Update patch set 2
Patch Set 2:

But you are introducing a third threshold. All three qqq messages only talk about the previously existing two thresholds. I understand that all three messages should be the same. This is fine. But it should be reworded. For example, it can talk about "softer" and "harder" thresholds, instead of the absolute "soft" and "hard".

Also you confused "to" with "too".

Patch-set: 2