'damaging' filter on Special:RC / Special:Watchlist
This filter is a not a typical show/hide toggle.
It is url-driven only for the moment and will be used
by the new RC filters (ERI project).

It accepts one or many of the following values:
  likelygood, maybebad, likelybad, verylikelybad
Values are separated by a comma.

Each value is associated with a range of probabilities
that are compared against the damaging test. They
can be configured using $wgOresDamagingLevels.

Bug: T149734
Change-Id: Ida6ab5e24a6da9487b10996f7edb6b9014c9184e