Remove default "return true" from all hook handlers
This is the default anyway. There is no difference between returning
true, null or nothing.

Note that I'm also removing a "return false" from the
OldChangesListRecentChangesLine hook! I believe this was misplaced.
Returning false means the recent changes line will be hidden. But why?
The failure does not mean the line is useless, it just means it will
not be colored.

Change-Id: Ic49f7819b2dd3c32943ee9149aa5db382fdcd03a