Update patch set 1
Patch Set 1:

IMO this is unnecessary except as a hack around the fact that the missing schema change on WMF wikis means that exceptional situations like job duplications can cause duplicate rows to be inserted into the tables. But I won't block it if people really want to do it anyway.

Patch-set: 1