Add ability to purge old rows to PurgeScoreCache
This allows PurgeScoreCache to be used as a cron job to limit
scores to those revisions which are still in recentchanges.

Each deletion batch does a left join on recentchanges which is
probably not the most efficient but seems good enough for a
maintenance script (P5319; 30ms/batch means 90M rows would
take roughly an hour).

Also add some logging per the TODO comment.

Bug: T159753
Change-Id: Id35bca820822dc46caa5adf6450c81871465abc9