Don't break the entire preferences page if the ORES API is down
If the ORES API is down or otherwise not giving us a response we expect,
and all damaging thresholds are defined in terms of model stats,
getDamagingThresholds() will fail to resolve all of them and return
an empty array. This caused the set of allowed values for the oresDamagingPref
preference to be set to an empty array, which then causes the
preferences page to die with an exception  because the default
value is not in the (empty) array of allowed values.

To fix this, don't make the allowed values for this preference
depend on the API result, but only on the $wgOresFiltersThresholds
config.

Change-Id: I94f551bb25aa20336a5ae93ed45d0b090bacf44f