Gracefully handle fetch stats errors
The test_stats ORES endpoint has been rather unstable lately.
In an effort to make the ORES-dependant pages degrade more
gracefully, this change logs the error and returns no levels
for the requested model. This will temporarily disable ORES
filters instead of crashing the page.

Also introducing $wgOresCacheVersion so the stats cache can
be invalidated when new stats are needed.

Change-Id: Icebb7f0dbb0a7cc44be801ccf7d136dfb07561c1