Use $wgOresRevisionsPerBatch == 0 to disable on-demand score fetch
The MediaWiki API forwards scoring requests to the ORES API when
those scores are not found in the local DB. To handle abuse
gracefully, this patch allows $wgOresRevisionsPerBatch = 0 to be
set to disable on-demand API lookups. Scores for new edits will
still be fetched but existing revisions will simply be returned
without a score if one is not cached already.

Also add a batch size parameter to PopulateDatabase so it's not
dependent on $wgOresRevisionsPerBatch being set.

Bug: T163687
Change-Id: Icf6f2a8760160f2a1122481759befa3fb1b318c2