Don't throw exception when "the" damaging threshold (for old code) can't be found
If the ORES service is down, or either it or this extension is misconfigured,
getDamagingThresholds() might return an empty array, or more generally,
an array that doesn't include the user's threshold preference.
In that case, have getThreshold() return null instead of throwing
an exception, and make sure callers handle null gracefully.

Bug: T163873
Change-Id: Ie771c1f5ce14b3dc4ade26392a0f0f416d6442df