Update patch set 1
Patch Set 1:

Putting this in because I think it will likely be helpful, but it's more of a proposal at this point. I think it might help with some slow queries looking for uncommon scores, but it's hard for me to predict how query behavior will change when an index is added. It would certainly help for things like T163040, if we ever build a feature like that.

I'm sure this should be linked to a phab task somewhere, but I got lost in the various adjacent tasks and didn't know what to link it to.

Patch-set: 1
Reviewer: Catrope <5@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0