Update patch set 1
Patch Set 1: Code-Review-1

I want to merge this as is, but the requirements have changed slightly (sorry!).  See my change to the bug description...  Now I'm thinking that legacy and new RC filters should be turned together, that our goal is to support deployment of the extension in a pure infrastructure mode.

If you agree, we should also suppress the preferences, the Special:Contributions controls, and legacy ORES for beta-decliners or noscripters.

Patch-set: 1
Label: Code-Review=-1