Update patch set 3
Patch Set 3:

One minor thing we may or may not care about: the entry for the "r" in the RC legend doesn't go away when you set $wgOresUiEnabled = false; . It's added through a config var directly in extension.json, so if we wanted to fix this we'd have to add an init callback and conditionally set it there. I considered doing that for other reasons (user may have prefs set such that "r" will never appear) but not sure if it's worth the effort.

Patch-set: 3