Hide RC/WL related preferences as appropriate
If the new filters are enabled on the RC page, hide the RC-related
preferences. This was already being done, but the logic didn't
take the new opt-out preference into account.

If the new filters are enabled on the Watchlist, hide the WL-related
preferences too, except for the ones that also affect Contributions.

Bug: T175611
Change-Id: Ia9355fb95e82c72185d2df93d48115f718fabe99