Add missing GPL file headers
This extension is licensed as GPL v3, or any later version (at your
option), so add the corresponding GPL file headers.

ApiQueryORES already had a GPL v2 or later file header, I updated it to
the v3 or later version for consistency with the rest of the extension.

WatchedItemQueryServiceExtension and ApiHooksHandler both had @author
tags for Brad, so I noted him as the copyright holder for both files.

Change-Id: Ie16893dbbe39fbd67b08f9da3d1e6dd3c69e454c