Update patch set 6
Patch Set 6:

For position in RC section, the ticket wants "Above Pending changes section, if possible.". Both prefs are added via Hooks and preferences array could be manipulated to adhere to preferences order, but I didn't find it necessary enough to increase complexity.

Also, there seem to be inconsistencies in how old and new RCFilter systems highlight the bad edits. If that needs to be consolidated as well, we might want new ticket for that.

Patch-set: 6
Label: Verified=0