Update patch set 4
Patch Set 4: Code-Review+1

(1 comment)

A -1? For what? I seriously can not see what should be wrong with this code. It's a small, easy to understand, nicely separated top-level factory. Sure, if there is an existing infrastructure that is equally small, separated and easy to understand, the code could also reuse that. But such a refactoring can be done any time and should not be a reason to block this patch from being merged.

Patch-set: 4
Label: Code-Review=+1