Update patch set 3
Patch Set 3: Code-Review-1

(1 comment)

I don't think making this checkbox sticky is the right way to solve this problem. The fundamental problem is that the checkboxes on the contributions page don't send &foo=0 when they're unchecked, they just drop out of the URL completely. The way the old Watchlist UI deals with this was by making sure that the checkbox is checked by default if the preference is set (it looks like ContributionsHooksHandler already does that), then when deciding whether the filter should be enabled or not, do the following:

- If action=submit is set: filter is enabled if the parameter is set, disabled otherwise
- If action=submit is not set: filter is enabled if the parameter is set or the preference is enabled, disabled otherwise

In other words, if action=submit is set, you ignore the preference, because you know it was already taken into account when setting the default state of the checkbox.

Patch-set: 3
Reviewer: Catrope <5@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1