Update patch set 2
Patch Set 2: Code-Review+1

(1 comment)

Typically a class *is* not what it says. It *represents* something. I like the example given: we don't name database classes "connectors", but "database".

I find the suggested name much better than the old one and are willing to merge this patch as it is. However, since ORES is an acronym for ObjectiveRevisionEvaluationService, the suggested name would almost literally end with "…ServiceService". Why not call it ObjectiveRevisionEvaluationService, and short $service when used as a variable?

Patch-set: 2
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1