Use SPDX 3.0 license identifier
SPDX released version 3 of their license list (<https://spdx.org/licenses/>),
which changed the FSF licenses to explicitly end in -only or -or-later
instead of relying on an easy to miss + symbol.

Also remove license info from package.json and set private: true
since we're not trying to publish this as a package.

Bug: T183858
Change-Id: Id744a9a96fe34e1f8c7dce488fda7989756529d7