Disable RCFilters in tests
RCFilters used to be disabled by default in Core and
enabled everywhere in config. Now the config has been
cleaned up and the feature is enabled by default in Core.

Setting this user option in the test re-creates the conditions
where the rcfilters features is disabled and ORES should
output the classic-rc highlighting module.

This is blocking merges of extensions depdending on
ORES (PageTriage).

Change-Id: I02a069006ef48ab3196ba7df7bf236e0f04f2b93