Fix broken condition in CleanDuplicateScores script
Note I have *no* idea if this is correct or not. I tried to understand
the code as good as I can, but I have never seen code using
buildGroupConcatField before. I think the idea is to find duplicates
(obviously, hence the name of the script). If the array contains more
then 1 hit, all but the 1st (array_slice( …, 1 ) cuts off the 1st) are
considered duplicates.

Comparing an array with "> 1" is *always* true, even for an empty array.

I think this was never a problem because $newIds is just an empty array
then. Which means the entire line with the broken "if ( … )" can as well
be removed.

Feel free to edit this patch as you like!

Change-Id: I01f9fa4ea0156a889e45e055ff42be8a185342c3