ext.ores.api: Single function
This removes most of the legacy copied from mw.Api which was
one of:

- specific to MW (e.g. '?format=json'),
- specific to needs of UploadWizard,
- specific to supporting multiple wikis,
- old quirks we regret, like https://phabricator.wikimedia.org/T176704.

It leaves a short wrapper around $.ajax() that sets a url, a timeout,
and ensures the data is parsed as JSON before being returned.

It also keeps the more or less random mw.log call we added in core
for some point for non-aborted request to be logged to the console,
although we may want to remove this as welll, given most browsers
now have superior debugging abilities enabled by default.

Change-Id: I40d8cd6caad7aca3e871892d49c6acd59c05ec5e