ThresholdLookup: Use a more descriptive key class for WANObjectCache
The top-level segment of makeKey() should be the logical grouping
of this kind of key (with variants as further segments). A good
rule of thumb is that any literals probably belong together
as the first segment.

This also makes the statistics provided by WANObjectCache more useful,
by avoiding unrelated operations from being tracked together as
"ORES". Instead, providing them for "ores-threshold-statistics".

Change-Id: Ic446bbb55936298197f1d195eec0847c9ccb4d82