Include filters config hash in cache key
Filters are configured using formulas that are resolved to
numerical thresholds using the ORES API. The API response
is cached per model for 24 hours.

When a model configuration changes slightly and its version
does not, the data in cache may not contain the result for
a specific formula and a filter level may become unusable.

This change proposes to include the md5 hash of the model config
in the cache key so that when it changes, the key is invalidated
and the model may be re-configured and re-cached immediately.

Change-Id: Idf89dd6669fe5edcecc7b9ddc69212b1587c4c89