Let the special page factory construct the SpecialPages
See I40fc8a709de96aceee14a10c973cd5b0a9a6f063, which injects
services into SpecialWatchList and SpecialRecentChanges.
The consequence of that change is that manually constructing
will causes errors, which are avoided by delegating this to
the SpecialPageFactory, which constructs the special pages with
the services injected.

Change-Id: Ia85c94216b573b64d971dea5bc92f6f3fea65eb9