feat: use Lift Wing instead of ORES
This patch uses Lift Wing to fetch scores instead of ORES.
The approach taken involves adding a LiftWingService in the ORESServices and extending ORESService.
We can switch ORES with LiftWing by setting the global config variable to OresUseLiftwing to true.
This patch also adapts some of the ORES code in order to make it extensible e.g. changing private to protected
attributes.
It is configured to use Lift Wing internal endpoint so the hostnames for the revscoring models that are required
to be in the header of the request are mapped withing the LiftWingService.
At the moment ORES service is still being used to fetch model thresholds, something that will be tackled in
a subsequent patch.

Bug: T319170
Change-Id: I351f4a60b959713b6cddd530af1c21db7e0bc1d6