feat: hardcode threshold calls to switch to Lift Wing
ORES supports querying the models and balance specificity with sensitivity based on queries provided
in the configuration for each model. In order to facilitate the transition to Lift Wing we load these thresholds from files
instead of fetching them from ORES. This is done by using a feature flag (global $wgOresUseLiftwing) to switch switching the ThresholdLookup service with the new service named
ThresholdLookupConfig.
In this patch we also fix the DumpThresholds.php maintenance script which was throwing an exception as it was trying
to access data in a wrong way (it was missing the wikiID key).

Bug: T319170
Change-Id: I0cd1e037e1e0217fcb914d9addfd67f654ca7bce