Add revertrisk model to the list of models
This patch adds revertrisk language agnostic to the list of models
available via the extension in recent changes page.

The model is not enabled by default.

We follow the same process that is followed for the ORES models.
The extension makes multiple requests to Lift Wing and then merges
them into one final response. For this reason we manipulate the
response returned by revertrisk
to match an ORES response.

Bug: T348298
Change-Id: Ib3f5460f940afb0a12b86c5fa3132ece309619a3