Change default config values to support local/patchdemo deployments
Change some default values to support out of the box deployments on localhost or
on patchdemo.
This involves the following changes:
- OresUseLiftwing: set default to true. This means that Lift Wing is being used
by default instead of ORES. Has no implications to current deployments
(production, beta etc) as all instances use Lift Wing already. This is mostly
set to help when developing locally of with patchdemo.

- OresLiftWingBaseUrl: Use the api gateway as the default url which is required
for external requests. Has no implication on production as the envoy services
proxy is defined in mediawiki-config (localhost:6031).

- OresLiftWingAddHostHeader: setting this to false. Requests made from the
extension to the API Gateway using the HttpRequestFactory will fail if a host header
 is provided. Switching this to false as it is turned on by default in
 mediawiki-config for production deployments.

Bug: T351703
Depends-On: I6669022d0f167b0da26510bbbe6e426da67fd49a
Change-Id: I612a84799afa8f8814ad32f8b921a67cb331d78d