SqlScoreStorage: round probabilities to nearest thousandth
The SQL field is a NUMERIC(3,3), so trying to insert
something with more precision, such as 0.123456 instead of
0.123, causes errors. 0.123456 will now be rounded to 0.123.

CI is currently broken for ORES and PageTriage. This patch
will fix CI.

The CI breakage is because rMW551ec29ea676 made this error
a bit louder, writing it to one of the logs that Jenkins
checks.

Bug: T355089

Change-Id: If043a46f198708420174943e1ae5be46c9d32b6e