Replace expensive explode/implode with string manipulation
There is no array here, only a single string with a single position
where we want to insert something. I found the old code quite
confusing esp. because it was doing everything multiple times.

I made sure this is well covered with tests. The new code does the
same as before.

Change-Id: I7fde3e16f368c79cabebe261fc38f249d509b01f