Improve filter group tests
My main goal was to remove the call to ChangesListSpecialPage::
getFilterGroups(), since I want to remove that method. But the RCL test
which called it was marked broken since 2019. That test was apparently
not testing anything, since it checked for no changes due to a hook
call, but the hook was already called by registerFilters(). The second
hook call would be a no-op regardless of the details.

Perhaps the test was trying to verify that there are no ORES filter
groups on RCL, but there are.

The reason for marking it broken was an interaction with Wikibase, which
I couldn't reproduce.

So:

* Factor out special page construction
* Inject a hook container into the special page with only the hook under
  test.
* Call setup() instead of registerFilters() since setup() is public.
* Rewrite the RCL test to document the current situation.

Bug: T224672
Bug: T403142
Change-Id: I757ae5f065672ec013f47635665bdd4f5ccb564a