ORES exception handling review
Errors can broadly be categorized as recoverable or unrecoverable.
Recoverable exceptions should almost always be caught. Unrecoverable
exceptions should never be caught.

For example, Excimer request timeouts, programmer errors and
configuration errors are unrecoverable.

Errors from HTTP APIs are generally recoverable, whereas database errors
are not. The difference stems from the way MW core handles database
transactions.

It is uncommon in the MediaWiki ecosystem to use exceptions to convey
recoverable errors: more often a Status or a special return value is
used. But it is an acceptable style, as long as it is done rigorously.

With this style of error handling, any time you catch an exception,
you need to add an exception class so that only the intended exceptions
are caught.

Out of frustration with uncaught errors, in T197647, catch blocks were
added around large blocks of ORES code, discarding all exceptions that
occurred within the blocks. This is not acceptable because unrecoverable
errors were discarded along with recoverable ones.

So:

* Activate the Phan mode which enforces rigorous use of @throws. This
  had a high true positive rate, detecting several bugs. False
  positives were mostly dealt with using
  exception_classes_with_optional_throws_phpdoc.
* Do not use @throws for unrecoverable errors.
* Use InvalidArgumentException only for unrecoverable programmer or
  configuration errors.
* Define exception classes for the caught exceptions thrown by the
  various backends.
* Instead of joinWithOresTables(), have maybeJoinWithOresTables(), which
  may or may not add a join depending on whether an exception was
  caught.
* Factor out duplicated code between WatchedItemQueryServiceExtension
  and ApiHooksHandler, and fix the error handling bugs in it.

Of the five bugs referenced in T197647, three involved
fetchThresholdsFromCache(), and two involved joinWithOresTables(). I'm
confident that there is no regression, because those two methods are
statically confirmed to be non-throwing now.

Change-Id: I861cd84d8566ffab490deb1ed4f8682648687424